chapter_001/src/test/java/ru/job4j/array/FindLoopTwoTest.java
package ru.job4j.array;

import org.junit.Test;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class FindLoopTwoTest {

    @Test
    public void whenArrayHasLengh5Then0() {
        FindLoopTwo find = new FindLoopTwo();
        int[] input = new int[] {5, 10, 3};
        int value = 5;
        int result = find.indexOf(input, value);
        int expect = 0;
        assertThat(result, is(expect));
    }

    @Test
    public void whenFind3() {
        FindLoopTwo find = new FindLoopTwo();
        int[] input = new int[] {5, 2, 10, 2, 4};
        int value = 2;
        int start = 2;
        int finish = 4;
        int result = find.indexOf(input, value, start, finish);
        int expect = 3;
        assertThat(result, is(expect));
    }

    @Test
    public void whenFindNo() {
        FindLoopTwo find = new FindLoopTwo();
        int[] input = new int[] {5, 2, 10, 2, 4};
        int value = 1;
        int start = 2;
        int finish = 4;
        int result = find.indexOf(input, value, start, finish);
        int expect = -1;
        assertThat(result, is(expect));
    }
}